[FLINK-16633][AZP] Fix builds without s3 credentials
Problem: Builds on Azure where failing if S3 credentails were not provided, because the syntax in the pipeline definition was incorrect.

Solution: Use a way to access secret variables that leads to an empty value (instead of the variable name). Also, extend the check in the unit tests to "variable empty" instead of just "variable not set".